+++ server/routes/emails.ts
@@ -2,49 +2,91 @@ import { Router } from 'express'
 import * as db from '../db/emails'
 import { logError } from '../logger'
 import { newEmailSchema } from '../../models/Emails'
+import { validateAccessToken } from '../auth0'
 import { isUserAdmin } from '../db/users'
-import { authorizeAdmin } from '../adminAuthorization'
-const adminUserId = 'auth0|def67890'
-const userId = 'auth0|abc12345'
 
 const router = Router()
 
-router.get('/today', authorizeAdmin, async (req, res) => {
+//!ADMIN ONLY
+router.get('/today', validateAccessToken, async (req, res) => {
   const today = new Date().toISOString().slice(0, 10)
 
-  try {
-    const unreadNumber = await db.getAmountOfUnreadEmailsByDate(today)
-    res.status(200).json(unreadNumber)
-  } catch (error) {
-    logError(error)
-    res.status(500).json({ message: 'failed to get number of unread emails' })
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
+
+  if (await isUserAdmin(userId)) {
+    try {
+      const unreadNumber = await db.getAmountOfUnreadEmailsByDate(today)
+      res.status(200).json(unreadNumber)
+    } catch (error) {
+      logError(error)
+      res.status(500).json({ message: 'failed to get number of unread emails' })
+    }
+  } else {
+    res.status(401).json({ message: 'user is not authorized as admin' })
   }
 })
 
-router.get('/', authorizeAdmin, async (req, res) => {
-  try {
-    const emails = await db.getAllEmails()
-    res.status(200).json({ emails })
-  } catch (error) {
-    logError(error)
-    res.status(500).json({ message: 'failed to get emails data from database' })
+//!ADMIN ONLY
+router.get('/', validateAccessToken, async (req, res) => {
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
+
+  if (await isUserAdmin(userId)) {
+    try {
+      const emails = await db.getAllEmails()
+      res.status(200).json({ emails })
+    } catch (error) {
+      logError(error)
+      res
+        .status(500)
+        .json({ message: 'failed to get emails data from database' })
+    }
+  } else {
+    res.status(401).json({ message: 'user is not authorized as admin' })
   }
 })
 
-router.get('/:id', authorizeAdmin, async (req, res) => {
-  // check auth0_id, if it has admin access
+//!ADMIN ONLY
+router.get('/:id', validateAccessToken, async (req, res) => {
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
   const emailId = Number(req.params.id)
-  try {
-    const email = await db.getEmailById(emailId)
-    res.status(200).json({ email })
-  } catch (error) {
-    logError(error)
-    res.status(500).json({ message: 'failed to get emails data from database' })
+  if (await isUserAdmin(userId)) {
+    try {
+      const email = await db.getEmailById(emailId)
+      res.status(200).json({ email })
+    } catch (error) {
+      logError(error)
+      res
+        .status(500)
+        .json({ message: 'failed to get emails data from database' })
+    }
+  } else {
+    res.status(401).json({ message: 'user is not authorized as admin' })
   }
 })
 
-router.post('/', async (req, res) => {
+router.post('/', validateAccessToken, async (req, res) => {
   const newEmail = req.body
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
 
   try {
     const result = newEmailSchema.safeParse(newEmail)
@@ -62,27 +104,49 @@ router.post('/', async (req, res) => {
   }
 })
 
-router.patch('/:id', authorizeAdmin, async (req, res) => {
-  try {
-    const id = Number(req.params.id)
-    await db.updateEmailReadStatusById(id)
-    res.sendStatus(200)
-  } catch (error) {
-    logError(error)
-    res
-      .status(500)
-      .json({ message: 'Unable to update read status to database' })
+//!ADMIN ONLY
+router.patch('/:id', validateAccessToken, async (req, res) => {
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
+  if (await isUserAdmin(userId)) {
+    try {
+      const id = Number(req.params.id)
+      await db.updateEmailReadStatusById(id)
+      res.sendStatus(200)
+    } catch (error) {
+      logError(error)
+      res
+        .status(500)
+        .json({ message: 'Unable to update read status to database' })
+    }
+  } else {
+    res.status(401).json({ message: 'user is not authorized as admin' })
   }
 })
 
-router.delete('/:id', authorizeAdmin, async (req, res) => {
-  try {
-    const id = Number(req.params.id)
-    await db.deleteEmailById(id)
-    res.sendStatus(200)
-  } catch (error) {
-    logError(error)
-    res.status(500).json({ message: 'could not delete email' })
+//!ADMIN ONLY
+router.delete('/:id', validateAccessToken, async (req, res) => {
+  const userId = req.auth?.payload.sub
+
+  if (!userId) {
+    res.status(400).json({ message: 'Please provide an id' })
+    return
+  }
+  if (await isUserAdmin(userId)) {
+    try {
+      const id = Number(req.params.id)
+      await db.deleteEmailById(id)
+      res.sendStatus(200)
+    } catch (error) {
+      logError(error)
+      res.status(500).json({ message: 'could not delete email' })
+    }
+  } else {
+    res.status(401).json({ message: 'user is not authorized as admin' })
   }
 })
